Suppressed some conversion warnings with C style casts. JIRA CXX-8578
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@76238 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/grid/test/test_netschedule_stress.cpp
@@ -130,7 +130,7 @@ void TestBatchSubmit(const string& service,
 
     NcbiCout << NcbiEndl << "Done." << NcbiEndl;
     double elapsed = sw.Elapsed();
-    double rate = jobs.size() / elapsed;
+    double rate = (double)jobs.size() / elapsed;
 
     NcbiCout.setf(IOS_BASE::fixed, IOS_BASE::floatfield);
     NcbiCout << "Elapsed: "  << elapsed << " sec." << NcbiEndl;
